+++ src/apig_wsgi/compat.py
@@ -8,18 +8,15 @@ if sys.version_info >= (3, 11):
     from wsgiref.types import WSGIApplication
 else:
     # Partial backport of wsgiref.types
+    from collections.abc import Iterable
     from types import TracebackType
     from typing import Any
     from typing import Callable
-    from typing import Dict
-    from typing import Iterable
     from typing import Protocol
-    from typing import Tuple
-    from typing import Type
     from typing import Union
 
-    _ExcInfo = Tuple[Type[BaseException], BaseException, TracebackType]
-    _OptExcInfo = Union[_ExcInfo, Tuple[None, None, None]]
+    _ExcInfo = tuple[type[BaseException], BaseException, TracebackType]
+    _OptExcInfo = Union[_ExcInfo, tuple[None, None, None]]
 
     class StartResponse(Protocol):
         """start_response() callable as defined in PEP 3333"""
@@ -32,5 +29,5 @@ else:
             # /,
         ) -> Callable[[bytes], object]: ...  # pragma: no cover
 
-    WSGIEnvironment = Dict[str, Any]
+    WSGIEnvironment = dict[str, Any]
     WSGIApplication = Callable[[WSGIEnvironment, StartResponse], Iterable[bytes]]